requirements.txt
aiohappyeyeballs==2.4.3 ; python_version >= "3.11" and python_version < "4.0"
aiohttp==3.10.10 ; python_version >= "3.11" and python_version < "4.0"
aiosignal==1.3.1 ; python_version >= "3.11" and python_version < "4.0"
altair==5.4.1 ; python_version >= "3.11" and python_version < "4.0"
annotated-types==0.7.0 ; python_version >= "3.11" and python_version < "4.0"
anyio==4.6.0 ; python_version >= "3.11" and python_version < "4.0"
attrs==24.2.0 ; python_version >= "3.11" and python_version < "4.0"
blinker==1.8.2 ; python_version >= "3.11" and python_version < "4.0"
cachetools==5.5.0 ; python_version >= "3.11" and python_version < "4.0"
certifi==2024.8.30 ; python_version >= "3.11" and python_version < "4.0"
cfgv==3.4.0 ; python_version >= "3.11" and python_version < "4.0"
charset-normalizer==3.4.0 ; python_version >= "3.11" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.11" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.11" and python_version < "4.0" and (sys_platform == "win32" or platform_system == "Windows")
distlib==0.3.9 ; python_version >= "3.11" and python_version < "4.0"
distro==1.9.0 ; python_version >= "3.11" and python_version < "4.0"
docstring-parser==0.16 ; python_version >= "3.11" and python_version < "4.0"
filelock==3.16.1 ; python_version >= "3.11" and python_version < "4.0"
frozenlist==1.4.1 ; python_version >= "3.11" and python_version < "4.0"
gitdb==4.0.11 ; python_version >= "3.11" and python_version < "4.0"
gitpython==3.1.43 ; python_version >= "3.11" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.11" and python_version < "4.0"
httpcore==1.0.6 ; python_version >= "3.11" and python_version < "4.0"
httpx==0.27.2 ; python_version >= "3.11" and python_version < "4.0"
identify==2.6.1 ; python_version >= "3.11" and python_version < "4.0"
idna==3.10 ; python_version >= "3.11" and python_version < "4.0"
iniconfig==2.0.0 ; python_version >= "3.11" and python_version < "4.0"
instructor==1.3.7 ; python_version >= "3.11" and python_version < "4.0"
jinja2==3.1.4 ; python_version >= "3.11" and python_version < "4.0"
jiter==0.4.2 ; python_version >= "3.11" and python_version < "4.0"
jsonschema-specifications==2024.10.1 ; python_version >= "3.11" and python_version < "4.0"
jsonschema==4.23.0 ; python_version >= "3.11" and python_version < "4.0"
markdown-it-py==3.0.0 ; python_version >= "3.11" and python_version < "4.0"
markupsafe==3.0.1 ; python_version >= "3.11" and python_version < "4.0"
mdurl==0.1.2 ; python_version >= "3.11" and python_version < "4.0"
multidict==6.1.0 ; python_version >= "3.11" and python_version < "4.0"
narwhals==1.9.3 ; python_version >= "3.11" and python_version < "4.0"
nodeenv==1.9.1 ; python_version >= "3.11" and python_version < "4.0"
numpy==2.1.2 ; python_version >= "3.11" and python_version < "4.0"
openai==1.33.0 ; python_version >= "3.11" and python_version < "4.0"
packaging==24.1 ; python_version >= "3.11" and python_version < "4.0"
pandas==2.2.3 ; python_version >= "3.11" and python_version < "4.0"
pillow==10.4.0 ; python_version >= "3.11" and python_version < "4.0"
platformdirs==4.3.6 ; python_version >= "3.11" and python_version < "4.0"
pluggy==1.5.0 ; python_version >= "3.11" and python_version < "4.0"
pre-commit==4.0.1 ; python_version >= "3.11" and python_version < "4.0"
propcache==0.2.0 ; python_version >= "3.11" and python_version < "4.0"
protobuf==5.28.2 ; python_version >= "3.11" and python_version < "4.0"
pyarrow==17.0.0 ; python_version >= "3.11" and python_version < "4.0"
pydantic-core==2.23.4 ; python_version >= "3.11" and python_version < "4.0"
pydantic==2.9.2 ; python_version >= "3.11" and python_version < "4.0"
pydeck==0.9.1 ; python_version >= "3.11" and python_version < "4.0"
pygments==2.18.0 ; python_version >= "3.11" and python_version < "4.0"
pytest==8.3.3 ; python_version >= "3.11" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.11" and python_version < "4.0"
python-dotenv==1.0.1 ; python_version >= "3.11" and python_version < "4.0"
pytz==2024.2 ; python_version >= "3.11" and python_version < "4.0"
pyyaml==6.0.2 ; python_version >= "3.11" and python_version < "4.0"
referencing==0.35.1 ; python_version >= "3.11" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.11" and python_version < "4.0"
rich==13.9.2 ; python_version >= "3.11" and python_version < "4.0"
rpds-py==0.20.0 ; python_version >= "3.11" and python_version < "4.0"
shellingham==1.5.4 ; python_version >= "3.11" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.11" and python_version < "4.0"
smmap==5.0.1 ; python_version >= "3.11" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.11" and python_version < "4.0"
streamlit==1.39.0 ; python_version >= "3.11" and python_version < "4.0"
swarm @ git+https://github.com/openai/swarm.git@d46acab2d41f4d2d351fa9a34fe0677da20f6296 ; python_version >= "3.11" and python_version < "4.0"
tenacity==8.5.0 ; python_version >= "3.11" and python_version < "4.0"
toml==0.10.2 ; python_version >= "3.11" and python_version < "4.0"
tornado==6.4.1 ; python_version >= "3.11" and python_version < "4.0"
tqdm==4.66.5 ; python_version >= "3.11" and python_version < "4.0"
typer==0.12.5 ; python_version >= "3.11" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version >= "3.11" and python_version < "4.0"
tzdata==2024.2 ; python_version >= "3.11" and python_version < "4.0"
urllib3==2.2.3 ; python_version >= "3.11" and python_version < "4.0"
virtualenv==20.26.6 ; python_version >= "3.11" and python_version < "4.0"
watchdog==5.0.3 ; python_version >= "3.11" and python_version < "4.0" and platform_system != "Darwin"
yarl==1.15.0 ; python_version >= "3.11" and python_version < "4.0"

